Obtain potential next wrapper from the builder
The builder should know about a potential next wrapper. GccBuilder tries
to detect ccache. ArBuilder does not allow any next wrapper.
