+++ pkg/common/constants.go
@@ -8,8 +8,6 @@ const (
 	QMSTRDEBUGENV = "QMSTR_DEBUG"
 	// CCACHEDIRENV is the name of the environment variable that stores the path to the ccache cache directory
 	CCACHEDIRENV = "CCACHE_DIR"
-	// QMSTRPREFIXENV is the name of the environment variable that defines a possible next compiler wrapper to invoke
-	QMSTRPREFIXENV = "QMSTR_PREFIX"
 )
 
 // ContainerBuildDir is where the source code gets mounted in the qmstr-master container
